+++ internal/sqlc/postgres/queries.sql
@@ -563,7 +563,7 @@ LIMIT 1;
 
 -- name: GetTypeGeneral :one
 SELECT
-    t.oid AS type_oid,
+    t.oid::int4 AS type_oid,
     t.typtype::text AS type_category,
     pg_get_userbyid(t.typowner) AS owner,
     COALESCE(obj_description(t.oid, 'pg_type'), '') AS comment
@@ -574,25 +574,25 @@ LIMIT 1;
 
 -- name: ListEnumLabels :many
 SELECT e.enumlabel FROM pg_enum e
-WHERE e.enumtypid = $1
+WHERE e.enumtypid::int4 = $1
 ORDER BY e.enumsortorder;
 
 -- name: ListCompositeAttributes :many
 SELECT a.attname, format_type(a.atttypid, a.atttypmod) AS data_type
 FROM pg_attribute a
-WHERE a.attrelid = (SELECT typrelid FROM pg_type WHERE oid = $1)
+WHERE a.attrelid = (SELECT typrelid FROM pg_type WHERE oid::int4 = $1)
   AND a.attnum > 0 AND NOT a.attisdropped
 ORDER BY a.attnum;
 
 -- name: GetRangeSubtype :one
 SELECT format_type(r.rngsubtype, NULL) AS subtype
 FROM pg_range r
-WHERE r.rngtypid = $1
+WHERE r.rngtypid::int4 = $1
 LIMIT 1;
 
 -- name: GetDomainGeneral :one
 SELECT
-    t.oid AS type_oid,
+    t.oid::int4 AS type_oid,
     format_type(t.typbasetype, t.typtypmod) AS base_type,
     t.typnotnull AS not_null,
     COALESCE(t.typdefault, '') AS default_value,
@@ -606,7 +606,7 @@ LIMIT 1;
 -- name: ListDomainConstraints :many
 SELECT c.conname, pg_get_constraintdef(c.oid, true) AS definition
 FROM pg_constraint c
-WHERE c.contypid = $1
+WHERE c.contypid::int4 = $1
 ORDER BY c.conname;
 
 -- name: ListCastsDetailed :many
